+++ test/ReadonlyArray.ts
@@ -35,6 +35,7 @@ import {
   extractAt,
   fromIterable,
   allM,
+  anyM,
 } from "../src/ReadonlyArray"
 import * as O from "fp-ts/Option"
 import * as A from "fp-ts/ReadonlyArray"
@@ -1120,6 +1121,37 @@ describe("Array", () => {
       expect(exe).toBe(false)
 
       f([IO.of(true), IO.of(true), set, IO.of(true)])()
+      expect(exe).toBe(true)
+    })
+    /* eslint-enable functional/no-expression-statement */
+  })
+
+  describe("anyM", () => {
+    const f = anyM(IO.Monad)
+
+    it("returns disjunctive identity on empty input", () => {
+      expect(f([])()).toBe(false)
+    })
+
+    it("equivalent to fold on lifted ||", () => {
+      expect(f([IO.of(true), IO.of(true)])()).toBe(true || true)
+      expect(f([IO.of(true), IO.of(false)])()).toBe(true || false)
+      expect(f([IO.of(false), IO.of(true)])()).toBe(false || true)
+      expect(f([IO.of(false), IO.of(false)])()).toBe(false || false)
+    })
+
+    /* eslint-disable functional/no-expression-statement */
+    it("runs from the left and short-circuits", () => {
+      let exe = false // eslint-disable-line functional/no-let
+      const set: IO<boolean> = () => {
+        exe = true
+        return true
+      }
+
+      f([IO.of(false), IO.of(true), set, IO.of(false)])()
+      expect(exe).toBe(false)
+
+      f([IO.of(false), IO.of(false), set, IO.of(false)])()
       expect(exe).toBe(true)
     })
     /* eslint-enable functional/no-expression-statement */